fix: adding elements to empty buffer
If adding (or pasting) an element into an otherwise empty DREX buffer
the 'load_buffer_content' "shortcut" in 'reload_directory' would
activate. After the content reload it would try to expand the root-path
within the root-path causing an error.

This was already mentioned in the comments, but not caught properly