fix: catch VIM error when opening files
For example when opening a file via 'vim.cmd(":e <file>")' which is
already opened in another VIM instance (swapfile exists) this will throw
the VIM error E325. The file opening will still happen, but an error
message will pop up afterwards which needs to be confirmed to continue.
With this fix we avoid this