fix(actions): fix focus element after renaming
If the target path after a rename is in scope of the current DREX buffer
and not expanded yet, the focus of the renamed element will be exactly
one line of. This happens because the focus function is called
immediately while the "old" element is still part of the current DREX
buffer. When it is deleted afterwards from its "old" location the cursor
will be shifted by a line

Therefore we now also monitor the former path for a change event (when
the element is removed from it) and call the focus function afterwards