refactor!: rename "is_valid_directory"
Rename the function to "points_to_existing_directory". This should
better clarify the purpose of this function and what is checked. Also it
should prevent confusion with the "is_(open_/closed_)directory"
functions and their completely different scope

This is breaking because the function was (and still is) publicly
usable. If you are currently using "is_valid_directory" in your
configuration you need to refactor it, otherwise this change does not
affect you